Updates Solid Queue dev config and Kamal deploy config.
Solid Queue now uses the DB instead of an in-memory queue. The reason for this is that recurring tasks (the scheduler) only work with the DB so we might as well have workers and dispatchers use it as well (see https://github.com/rails/solid_queue?tab=readme-ov-file#workers-dispatchers-and-scheduler).

I wanted to ensure Kamal deploy works so I followed these instructions to deploy to a local VM: https://alexpeattie.com/blog/testing-kamal-locally-with-multipass.
I'm just going to check in the config as is, which is configured to a certain IP, ssh user,  as well as my DockerHub account. For production we'd just need to update the IP and ssh user.